refactor(task-2): complete batch management page decomposition
MAJOR REFACTOR: 509 LOC → 191 LOC shell (62% reduction)

Decomposed into production-quality feature slice:
✅ Shell page: BatchManagementPage.tsx (191 LOC, orchestration only)
✅ Hooks: useBatchFilters, useBatchCreation (business logic)
✅ Utils: batchHelpers, lifecycleHelpers (pure functions)
✅ Schema: batchFormSchema (form validation)
✅ Component: CreateBatchDialog (dialog + form)
✅ Tests: 27 new unit tests (100% passing)

Quality metrics:
- Tests: 396 passed (+27 new) | 7 skipped
- Type checking: ✅ PASS (0 errors)
- Complexity: No new warnings (still 2 total)
- Total NLOC: 7,813 (+366 from utilities & tests)
- Avg CCN: 1.6 (unchanged, excellent)

Backward compatibility: Old path re-exports new location
Architecture: Clean separation - shell delegates to hooks
Production ready: Full error handling, JSDoc, tests
